+++ experiment/fast/src/lib.rs
@@ -1,16 +1,11 @@
-pub(crate) mod measure;
+mod measure;
 mod notation;
-#[cfg(test)]
-mod oracular_pretty_print;
-pub(crate) mod requirement;
+mod requirement;
 //mod partial_pretty_print;
 mod pretty_print;
-#[cfg(test)]
-mod random_notation;
-#[cfg(test)]
-mod random_requirement;
-mod staircase;
+//mod staircase;
 mod validate;
 
 pub use notation::Notation;
 pub use pretty_print::pretty_print;
+pub use requirement::{AlignedMultiLine, MultiLine, Requirement};
